+++ .gitignore
@@ -1,2 +1,3 @@
 /target
 .DS_Store
+.venv
