+++ aderyn/Cargo.toml
@@ -9,7 +9,5 @@ license = "MIT"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-aderyn_core = { path  = "../aderyn_core" }
+aderyn_driver = { path  = "../aderyn_driver" }
 clap = { version = "4.4.6", features = ["derive"] }
-rayon = "1.8.0"
-tokei = "12.1.2"
