+++ aderyn_driver/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "aderyn_driver"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+aderyn_core = { path  = "../aderyn_core" }
+rayon = "1.8.0"
+tokei = "12.1.2"
