+++ aderyn_driver/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod driver;
 pub mod process_foundry;
 pub mod process_hardhat;
 pub mod virtual_foundry;
